build the server using Akka Streams / Reactive Kafka and Akka HTTP, was not sure where to use Scala/Java so it became a weird mix, I have not verified everything works yet... if you can just run it if you have the kafka env for it etc that would be nice. I'll be around to help etc, but if we just want to share sample code in article those shouldlikely be good already.
Notice that we use prety advanced Akka Streams things - the stage has
state - the model, and ha 2 streaming inputs, the changes and the wine
data, outputs are for each scored wine, and one can inspect current
state from outside of the streaming thing by the materialized value (the
ReadableModelStateStore) similar to how it was previously.

Would be nice to not show jetty / servlets / jax but Akka HTTP in this
context - serving those over http. If you need this as Java let me know
and i'll convert it.

The ModelGenerator also was a good example how that can be pretty
consice with Akka Streams, so I have converted that too, feel free to
use or not, it could fit the story and explain the positioning what Akka
Streams are - not a "spark like thing" which is the most common
misconception I guess.